% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_history_range.R
\name{coin_history_range}
\alias{coin_history_range}
\title{Get historical market data}
\usage{
coin_history_range(coin_id, vs_currency, from, to, max_attempts = 3)
}
\arguments{
\item{coin_id}{(character): ID of the coin of interest. An
up-to-date list of supported coins and their IDs can be retrieved
with the \code{\link[=supported_coins]{supported_coins()}} function.}

\item{vs_currency}{(character): name of the base currency to
to express the price in. An up-to-date list of supported currencies (both
fiat and cryptocurrencies) can be retrieved with the \code{\link[=supported_currencies]{supported_currencies()}}
function. If an unsupported \code{vs_currency} is requested, the call will
fail with the respective error message.}

\item{from}{(POSIXct): timestamp of the beginning of the period of interest
(\verb{YYYY-MM-DD HH:MM:SS}).}

\item{to}{(POSIXct): timestamp of the end of the period of interest
(\verb{YYYY-MM-DD HH:MM:SS}).}

\item{max_attempts}{(double, positive): specifies the
maximum number of attempts to call the CoinGecko API (e.g., if
the first call fails, with a status different from \code{200}).
Additional attempts are implemented with an exponential backoff.
Defaults to 3.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{timestamp} (POSIXct);
\item \code{vs_currency} (character): same as the argument \code{vs_currency};
\item \code{price} (double): coin price as of \code{timestamp};
\item \code{total_trading_vol} (double): a 24 h rolling-window trading volume, as
of \code{timestampt};
\item \code{market_cap} (double): market capitalisation, as of \code{timestamp}.
}
}
\description{
Retrieves coin-specific market data for a range of historical dates
}
\details{
This function returns hourly data for periods of up to 90 days
and daily data for periods longer than 90 days.

This function is based on the public
CoinGecko API, which has a limit of 100 calls per minute. Please
keep this limit in mind when developing your applications.
}
\examples{
r <- coin_history_range(
  coin_id = "cardano",
  vs_currency = "usd",
  from = as.POSIXct("2021-01-01 13:00:00", tz = "UTC"),
  to = as.POSIXct("2021-01-01 18:00:00", tz = "UTC")
)
print(r)
}
