% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_price.R
\name{crypto_price}
\alias{crypto_price}
\title{Get the current cryptocurrency prices}
\usage{
crypto_price(
  coin_ids,
  vs_currencies = c("usd"),
  include_market_cap = TRUE,
  include_24h_vol = TRUE,
  include_24h_change = TRUE,
  max_attempts = 3L
)
}
\arguments{
\item{coin_ids}{(character): a vector with IDs of the cryptocurrencies of
interest. An up-to-date list of supported cryptocurrencies and their
IDs can be retrieved with the \code{\link[=supported_cryptos]{supported_cryptos()}}
function.}

\item{vs_currencies}{(character): a vector with
names of the base currencies to benchmark against, e.g.
\code{c("usd", "eur", "btc")}. An up-to-date list of supported
\code{vs_currencies} (both fiat and cryptocurrencies) can be retrieved with
the \code{\link[=supported_currencies]{supported_currencies()}} function. If \code{vs_currencies}
contains at least one unsupported currency, the call will fail with the
respective error message.}

\item{include_market_cap}{(boolean, defaults to
\code{TRUE}): whether to return the market cap information.}

\item{include_24h_vol}{(boolean, defaults to \code{TRUE}):
whether to return the trading volume for the last 24 hours.}

\item{include_24h_change}{(boolean, defaults to
\code{TRUE}): whether to return the price percentage change compared to 24
hours ago.}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A tibble, which by the default will have the following columns:
\itemize{
\item \code{coin_id} (character): IDs of the cryptocurrencies, ordered alphabetically;
\item \code{vs_currency} (character): names of the base currencies to express the
price in;
\item \code{last_updated_at} (POSIXct, UTC timezone): timestamp of the last price
update;
\item \code{market_cap} (double): current market cap;
\item \code{vol_24h} (double): trading volume in the last 24 hours;
\item \code{price_percent_change_24h} (double): percentage change of the price
compared to 24 hours ago.
}
}
\description{
The current price of any cryptocurrencies in any other supported currencies
}
\details{
This function is based on the public
CoinGecko API, which has a limit of 100 calls per minute. Please
keep this limit in mind when developing your applications.
}
\examples{
r <- crypto_price(coin_ids = c("aave", "tron", "btc"),
                  vs_currencies = c("usd", "eur", "gbp"))
print(r)
}
