% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_params.R
\name{function_params}
\alias{function_params}
\title{Definitions of common function arguments}
\usage{
function_params(arguments)
}
\arguments{
\item{arguments}{(character): vector with names of the arguments to be
documented.}
}
\value{
Literal descriptions of the arguments.
}
\description{
Used to automatically document frequently used arguments
}
\keyword{internal}
