% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_history_panel.R
\name{get_history_panel}
\alias{get_history_panel}
\title{Get historical data for panel of coins}
\usage{
get_history_panel(coin_ids, ...)
}
\arguments{
\item{coin_ids}{the list or vector containing the id's of the coins. This comes from using \code{geckor::supported_coins()}.}

\item{...}{additional parameters to be passed to \code{geckor::coin_history()}}

\item{maxcoins}{Limit the query to the first 'maxcoins' in the coin_ids list. To get the full list of coins, run \code{geckor::supported_coins()}}
}
\value{
a data table with a panel of coins at daily or more granular frequency. For details, see \code{geckor::coin_history()}.
}
\description{
Get historical data for panel of coins
}
