% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_cryptos.R
\name{supported_cryptos}
\alias{supported_cryptos}
\title{Get CoinGecko currencies}
\usage{
supported_cryptos(max_attempts = 3L)
}
\arguments{
\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A tibble with three columns:
\itemize{
\item \code{coin_id} (character): IDs of the cryptocurrencies;
\item \code{symbol} (character): abbreviated names of the cryptocurrencies;
\item \code{name} (character): common names of the cryptocurrencies.
}
}
\description{
Retrieves a list of cryptocurrencies currently supported by the CoinGecko API
}
\details{
This function is based on the public
CoinGecko API, which has a limit of 100 calls per minute. Please
keep this limit in mind when developing your applications.
}
\examples{
r <- supported_cryptos()
head(r, 10)
}
