% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_currencies.R
\name{supported_currencies}
\alias{supported_currencies}
\title{Get CoinGecko currencies}
\usage{
supported_currencies(max_attempts = 3)
}
\arguments{
\item{max_attempts}{(double, positive): specifies the
maximum number of additional attempts to call the CoinGecko API
if the first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to 3.}
}
\value{
Character vector with abbreviated names of the currencies.
}
\description{
Retrieves a list of base currencies currently supported by the CoinGecko API
}
\details{
This function is based on the public
CoinGecko API, which has a limit of 100 calls per minute. Please
keep this limit in mind when developing your applications.
}
\examples{
r <- supported_currencies()
print(r)
}
