% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_exchanges.R
\name{supported_exchanges}
\alias{supported_exchanges}
\title{Get supported exchanges}
\usage{
supported_exchanges(max_attempts = 3L)
}
\arguments{
\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
Character vector with abbreviated names of the exchanges.
}
\description{
Retrieves a list of exchanges supported by the CoinGecko API
}
\details{
This function is based on the public
CoinGecko API, which has a limit of 100 calls per minute. Please
keep this limit in mind when developing your applications.
}
\examples{
r <- supported_exchanges()
print(r)
}
