% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arguments.R
\name{validate_arguments}
\alias{validate_arguments}
\title{Validate arguments}
\usage{
validate_arguments(
  arg_max_attempts = NULL,
  arg_coin_ids = NULL,
  arg_vs_currencies = NULL,
  arg_exchange_id = NULL,
  arg_include_market_cap = NULL,
  arg_include_24h_vol = NULL,
  arg_include_24h_change = NULL,
  arg_date = NULL
)
}
\value{
This function is to be used for its side effects. If any of the
checks is not passed, the respective error is returned. If all checks
are passed successfully, nothing is returned.
}
\description{
Validates arguments that are frequently used in \code{geckor} functions
}
\keyword{internal}
