% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_market.R
\name{current_market}
\alias{current_market}
\title{Current market status}
\usage{
current_market(coin_ids, vs_currency, max_attempts = 3L)
}
\arguments{
\item{coin_ids}{(character): a vector with IDs of the coins of
interest. An up-to-date list of supported coins and their
IDs can be retrieved with the \code{\link[=supported_coins]{supported_coins()}} function.}

\item{vs_currency}{(character): name of the base currency to
benchmark against. An up-to-date list of supported currencies (both fiat
and cryptocurrencies) can be retrieved with the \code{\link[=supported_currencies]{supported_currencies()}}
function. If an unsupported \code{vs_currency} is requested, the call will
fail with the respective error message.}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
If the requested data exist, this function will return a tibble with
with as many rows as the length of \code{coin_ids} and the following columns:
\itemize{
\item \code{coin_id} (character): coin IDs, ordered by \code{market_cap} (see below);
\item \code{symbol} (character): coin symbol;
\item \code{name} (character): coin common name;
\item \code{last_updated_at} (POSIXct, UTC timezone): timestamp of the last update;
\item \code{current_price} (double): current price (as of \code{last_updated_at}),
expressed in \code{vs_currency};
\item \code{market_cap} (double): current market capitalisation;
\item \code{market_cap_rank} (integer): current rank of the coin in terms of its
market capitalisation;
\item \code{fully_diluted_valuation} (double): fully diluted valuation of the coin'
project;
\item \code{total_volume} (double): total trading volume in the last 24 hours;
\item \code{high_24h} (double): max price recorded in the last 24 hours;
\item \code{low_24h} (double): min price recorded in the last 24 hours;
\item \code{price_change_24h} (double): price change as compared to 24 hours ago;
\item \code{price_change_percentage_24h} (double): percentage change of the price as
compared to 24 hours ago;
\item \code{circulating_supply} (double): coin supply currently in circulation;
\item \code{total_supply} (double): total supply that can potentially be circulated;
\item \code{max_supply} (double): max possible supply that can be circulated;
\item \code{ath} (double): all-time high price;
\item \code{ath_change_percentage} (double): percentage change of the all-time high
price compared to the current price;
\item \code{ath_date} (POSIXct, UTC timezone): timestamp of when the all-time high
price has been recorded;
\item \code{atl} (double): all-time low price;
\item \code{atl_change_percentage} (double): percentage change of the all-time low
price compared to the current price;
\item \code{atl_date} (POSIXct, UTC timezone): timestamp of when the all-time low
price has been recorded;
\item \verb{price_change_percentage_<time>_in_currency}: columns containing the
percentage change in price as compared to various periods in the past
(in particular, 1 hour, 24 hours, 7 days, 14 days, 30 days, 200 days,
and 1 year).
}
}
\description{
Retrieves metrics describing the current market status of a set of coins
}
\details{
If no data can be retrieved (e.g., because of the misspecified
query parameters), this function will return nothing (\code{NULL}).

This function is based on the public
CoinGecko API, which has a limit of 100 calls per minute. Please
keep this limit in mind when developing your applications.
}
\examples{
r <- current_market(
  coin_ids = c("bitcoin", "ethereum", "cardano"),
  vs_currency = "usd"
)
print(r)
}
