% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_coins.R
\name{supported_coins}
\alias{supported_coins}
\title{Get CoinGecko coins}
\usage{
supported_coins(max_attempts = 3)
}
\arguments{
\item{max_attempts}{(double, positive): specifies the
maximum number of additional attempts to call the CoinGecko API
if the first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to 3.}
}
\value{
A tibble with three columns:
\itemize{
\item \code{coin_id}(character): coin IDs, ordered alphabetically;
\item \code{symbol} (character): coin symbol;
\item \code{name} (character): coin common name.
}
}
\description{
Retrieves a list of coins currently supported by the CoinGecko API
}
\details{
This function is based on the public
CoinGecko API, which has a limit of 100 calls per minute. Please
keep this limit in mind when developing your applications.
}
\examples{
\donttest{
r <- supported_coins()
head(r, 10)
}
}
