% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_currencies.R
\name{supported_currencies}
\alias{supported_currencies}
\title{Get CoinGecko currencies}
\usage{
supported_currencies(max_attempts = 3L)
}
\arguments{
\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
Character vector with abbreviated names of the currencies.
}
\description{
Retrieves a list of base currencies currently supported by the CoinGecko API
}
\examples{
r <- supported_currencies()
print(r)
}
