% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arguments.R
\name{validate_arguments}
\alias{validate_arguments}
\title{Validate arguments}
\usage{
validate_arguments(arg_max_attempts = NULL, arg_vs_currencies = NULL)
}
\arguments{
\item{arg_max_attempts}{(integer): number of additional attempts to call
the API if the first attempt fails.}

\item{arg_vs_currencies}{(character): abbreviated names of the currencies
to benchmark against.}
}
\value{
This function is to be used for its side effects. If any of the
checks is not passed, it returns the respective error. If all checks
are passed successfully, nothing is returned.
}
\description{
Validates arguments that are commonly used in \code{geckor} functions
}
\keyword{internal}
